Chart Dispatcher
### Description
Exposes an API at `<url>/chart?chat_id=<id>` and after generating a
basic Candle Stick Chart sends it through the bot api.

###To Do
The data provided are now sample, but the function `createChart` it’s
easily adjustable with call to the datacenter api in order to populate
a chart with proper values.